count:int -> long long

+++ sim/sim_core/type.h
@@ -117,7 +117,7 @@ void  print_label_of_pc(int pc,LList llist);
 // instr
 Instr    initialize_instr(void);
 void     exec_instr(Instr i, Mem memory, Reg reg);
-int      exec_program(Program program, Reg reg, Mem mem);
+long long      exec_program(Program program, Reg reg, Mem mem);
 
 // runtime
 Mem      initialize_memory(int memsize, Mem p);
